requirements.txt
autopep8==2.0.2
fastapi==0.95.0
flake8==6.0.0
gunicorn
matplotlib
pymongo
numpy
openai
psycopg2
pytest
python-dotenv
sqlalchemy
uvicorn[standard]==0.20.0
